feat: 랭킹과 리뷰에서 사용할 수 있게 variant 분리
